🏗️ Fix #1253: マネージャー統合完了 - 6個→3個への統合 (50%削減)
## Summary
- 6個のマネージャーを3個に統合（50%削減達成）
- 機能統合により管理複雑性を削減
- unified_api.pyとの互換性維持

## Changes
### 統合されたマネージャー構成:
1. **CoreManager** (拡張)
   - 旧CoreManager + ChunkManager + DistributionManager
   - IO・テンプレート・チャンク・配布管理を統合

2. **ProcessingManager** (新規)
   - 旧ParsingManager + OptimizationManager
   - 解析・バリデーション・最適化処理を統合

3. **PluginManager** (維持)
   - 独立性を保持してそのまま維持

### 削除されたファイル:
- kumihan_formatter/managers/parsing_manager.py
- kumihan_formatter/managers/optimization_manager.py
- kumihan_formatter/core/processing/chunk_manager.py
- kumihan_formatter/core/io/distribution_manager.py

### 修正されたファイル:
- kumihan_formatter/managers/__init__.py
- kumihan_formatter/managers/core_manager.py
- kumihan_formatter/core/api/manager_coordinator.py
- kumihan_formatter/managers/processing_manager.py (新規)

## Test Results
✅ 全マネージャーの初期化テスト成功
✅ Black + mypy品質チェック通過 (エラー40個 < 150個目標)
