+++ kumihan_formatter/managers/processing_manager.py
@@ -1,25 +1,47 @@
-"""
-ParsingManager - 解析機能統合管理クラス
-ParseManager + ValidationManager の機能を統合
-"""
-
+from typing import Any, Dict, List, Optional, Union, Callable
 import logging
-from typing import Any, Dict, List, Optional, Union
+import time
+import inspect
+from functools import wraps
 
-from kumihan_formatter.core.ast_nodes.node import Node
+"""
+ProcessingManager - 解析・最適化処理統合管理クラス (Issue #1253対応)
+ParsingManager + OptimizationManager の機能を統合
+"""
+
+# パーシング関連インポート
 from kumihan_formatter.core.processing.parsing_coordinator import ParsingCoordinator
+from kumihan_formatter.core.validation.validation_reporter import ValidationReporter
 from kumihan_formatter.parsers.unified_list_parser import UnifiedListParser
 from kumihan_formatter.parsers.unified_keyword_parser import UnifiedKeywordParser
 from kumihan_formatter.parsers.unified_markdown_parser import UnifiedMarkdownParser
-from kumihan_formatter.core.validation.validation_reporter import ValidationReporter
+from kumihan_formatter.core.ast_nodes.node import Node
 
 
-class ParsingManager:
-    """解析機能統合管理クラス - 高レベル解析・検証API"""
+class PerformanceMetrics:
+    """パフォーマンス測定メトリクス（OptimizationManager統合）"""
+
+    def __init__(
+        self,
+        operation_name: str,
+        execution_time: float,
+        memory_usage: int,
+        input_size: int,
+        optimization_applied: bool,
+    ):
+        self.operation_name = operation_name
+        self.execution_time = execution_time
+        self.memory_usage = memory_usage
+        self.input_size = input_size
+        self.optimization_applied = optimization_applied
+
+
+class ProcessingManager:
+    """解析・最適化処理統合管理クラス - パーシング・バリデーション・パフォーマンス最適化の統合API (Issue #1253対応)"""
 
     def __init__(self, config: Optional[Dict[str, Any]] = None):
         """
-        ParsingManager初期化
+        ProcessingManager初期化
 
         Args:
             config: 設定オプション辞書
@@ -27,7 +49,7 @@ class ParsingManager:
         self.logger = logging.getLogger(__name__)
         self.config = config or {}
 
-        # 既存コンポーネント初期化
+        # パーシング機能初期化（旧ParsingManager）
         self.coordinator = ParsingCoordinator(config)
         self.reporter = ValidationReporter()
 
@@ -40,7 +62,17 @@ class ParsingManager:
         self.strict_mode = self.config.get("strict_validation", False)
         self.error_threshold = self.config.get("error_threshold", 10)
 
-    # ========== パーシング機能（旧ParseManager） ==========
+        # 最適化設定（旧OptimizationManager）
+        self.enable_caching = self.config.get("enable_caching", True)
+        self.enable_parallel = self.config.get("enable_parallel", True)
+        self.memory_limit = self.config.get("memory_limit_mb", 512)
+        self.performance_monitoring = self.config.get("performance_monitoring", True)
+
+        # パフォーマンス測定
+        self._metrics: List[PerformanceMetrics] = []
+        self._operation_cache: Dict[str, Any] = {}
+
+    # ========== パーシング機能（旧ParsingManager） ==========
 
     def parse(
         self, content: Union[str, List[str]], parser_type: str = "auto"
@@ -116,7 +148,110 @@ class ParsingManager:
             self.logger.error(f"ファイル解析中にエラー: {file_path}, {e}")
             return None
 
-    # ========== バリデーション機能（旧ValidationManager） ==========
+    # ========== 最適化パーシング（OptimizationManager統合） ==========
+
+    def optimize_parsing(
+        self, content: Union[str, List[str]], parser_func: Callable[[str], Any]
+    ) -> Any:
+        """
+        パーシング処理の最適化
+
+        Args:
+            content: 解析対象コンテンツ
+            parser_func: パーサー関数
+
+        Returns:
+            最適化された解析結果
+        """
+        try:
+            start_time = time.time()
+
+            # コンテンツサイズチェック
+            if isinstance(content, str):
+                input_size = len(content)
+                content_hash = hash(content)
+            else:
+                input_size = sum(len(line) for line in content)
+                content_hash = hash(str(content))
+
+            # キャッシュチェック
+            cache_key = f"parse_{content_hash}_{parser_func.__name__}"
+            if self.enable_caching and cache_key in self._operation_cache:
+                self.logger.debug(f"キャッシュヒット: {parser_func.__name__}")
+                cached_result = self._operation_cache[cache_key]
+
+                # メトリクス記録
+                execution_time = time.time() - start_time
+                self._record_metrics(
+                    "parse_cached", execution_time, 0, input_size, True
+                )
+
+                return cached_result
+
+            # 最適化されたパーシング実行
+            content_str = content if isinstance(content, str) else "\n".join(content)
+            if input_size > 50000:  # 大きなコンテンツの場合
+                result = self._optimize_large_parsing(content_str, parser_func)
+            else:
+                result = parser_func(content_str)
+
+            # キャッシュ保存
+            if self.enable_caching:
+                self._operation_cache[cache_key] = result
+
+            # メトリクス記録
+            execution_time = time.time() - start_time
+            self._record_metrics(
+                parser_func.__name__,
+                execution_time,
+                self._estimate_memory_usage(result),
+                input_size,
+                True,
+            )
+
+            return result
+
+        except Exception as e:
+            self.logger.error(f"パーシング最適化中にエラー: {e}")
+            # フォールバック: 通常のパーシング
+            content_str = content if isinstance(content, str) else "\n".join(content)
+            return parser_func(content_str)
+
+    def _optimize_large_parsing(
+        self, content: Union[str, List[str]], parser_func: Callable[[str], Any]
+    ) -> Any:
+        """大きなコンテンツの最適化パーシング"""
+        try:
+            if isinstance(content, str):
+                lines = content.split("\n")
+            else:
+                lines = content
+
+            # チャンク分割による並列処理（簡易版）
+            chunk_size = self.config.get("large_parse_chunk_size", 1000)
+            chunks = [
+                lines[i : i + chunk_size] for i in range(0, len(lines), chunk_size)
+            ]
+
+            results = []
+            for chunk in chunks:
+                chunk_str = "\n".join(chunk)
+                chunk_result = parser_func(chunk_str)
+                if chunk_result:
+                    results.append(chunk_result)
+
+            # 結果統合（簡易版）
+            if results:
+                return results[0]  # 最初の有効な結果を使用
+
+            return None
+
+        except Exception as e:
+            self.logger.warning(f"大容量パーシング最適化に失敗、フォールバック: {e}")
+            content_str = content if isinstance(content, str) else "\n".join(content)
+            return parser_func(content_str)
+
+    # ========== バリデーション機能（旧ParsingManager） ==========
 
     def validate_node(
         self, node: Node, context: Optional[Dict[str, Any]] = None
@@ -319,6 +454,103 @@ class ParsingManager:
                         f"行{line_num}: 記法が曖昧です - {line.strip()[:50]}"
                     )
 
+    # ========== パフォーマンス監視（OptimizationManager統合） ==========
+
+    def performance_monitor(self, func: Callable[..., Any]) -> Callable[..., Any]:
+        """
+        関数実行パフォーマンス監視デコレーター
+
+        Args:
+            func: 監視対象関数
+
+        Returns:
+            監視付き関数
+        """
+
+        @wraps(func)
+        def wrapper(*args: Any, **kwargs: Any) -> Any:
+            if not self.performance_monitoring:
+                return func(*args, **kwargs)
+
+            start_time = time.time()
+            try:
+                result = func(*args, **kwargs)
+                execution_time = time.time() - start_time
+
+                # メトリクス記録
+                self._record_metrics(
+                    func.__name__,
+                    execution_time,
+                    self._estimate_memory_usage(result),
+                    self._estimate_input_size(args, kwargs),
+                    False,
+                )
+
+                return result
+
+            except Exception as e:
+                execution_time = time.time() - start_time
+                self.logger.error(f"監視対象関数でエラー ({func.__name__}): {e}")
+                self._record_metrics(func.__name__, execution_time, 0, 0, False)
+                raise
+
+        return wrapper
+
+    def _record_metrics(
+        self,
+        operation_name: str,
+        execution_time: float,
+        memory_usage: int,
+        input_size: int,
+        optimization_applied: bool,
+    ) -> None:
+        """パフォーマンスメトリクスの記録"""
+        metric = PerformanceMetrics(
+            operation_name=operation_name,
+            execution_time=execution_time,
+            memory_usage=memory_usage,
+            input_size=input_size,
+            optimization_applied=optimization_applied,
+        )
+
+        self._metrics.append(metric)
+
+        # メトリクス数制限（メモリ節約）
+        if len(self._metrics) > 1000:
+            self._metrics = self._metrics[-500:]  # 最新500件を保持
+
+    # ========== メモリ最適化 ==========
+
+    def optimize_memory_usage(self) -> Dict[str, Any]:
+        """
+        メモリ使用量最適化
+
+        Returns:
+            最適化結果
+        """
+        try:
+            initial_cache_size = len(self._operation_cache)
+
+            # 古いキャッシュエントリの削除（簡易版LRU）
+            if len(self._operation_cache) > 100:
+                # 最近使用されたもの以外を削除
+                cache_items = list(self._operation_cache.items())
+                recent_items = cache_items[-50:]  # 最新50項目を保持
+                self._operation_cache = dict(recent_items)
+
+            optimized_size = len(self._operation_cache)
+
+            return {
+                "cache_cleaned": True,
+                "initial_cache_size": initial_cache_size,
+                "optimized_cache_size": optimized_size,
+                "memory_freed": initial_cache_size - optimized_size,
+            }
+
+        except Exception as e:
+            self.logger.error(f"メモリ最適化中にエラー: {e}")
+            return {"cache_cleaned": False, "error": str(e)}
+
     # ========== 統合API ==========
 
     def parse_and_validate(
@@ -367,6 +599,8 @@ class ParsingManager:
                 "overall_valid": False,
             }
 
+    # ========== 情報取得 ==========
+
     def get_available_parsers(self) -> List[str]:
         """利用可能なパーサー一覧を取得"""
         return ["auto", "list", "keyword", "markdown"]
@@ -383,3 +617,56 @@ class ParsingManager:
             },
             "config": self.config,
         }
+
+    def get_optimization_statistics(self) -> Dict[str, Any]:
+        """最適化統計情報を取得"""
+        if not self._metrics:
+            return {"total_operations": 0, "optimization_rate": 0.0}
+
+        total_ops = len(self._metrics)
+        optimized_ops = sum(1 for m in self._metrics if m.optimization_applied)
+
+        avg_execution_time = sum(m.execution_time for m in self._metrics) / total_ops
+
+        return {
+            "total_operations": total_ops,
+            "optimized_operations": optimized_ops,
+            "optimization_rate": optimized_ops / total_ops,
+            "avg_execution_time": avg_execution_time,
+            "cache_size": len(self._operation_cache),
+            "config": {
+                "caching_enabled": self.enable_caching,
+                "parallel_enabled": self.enable_parallel,
+                "memory_limit_mb": self.memory_limit,
+            },
+        }
+
+    def clear_optimization_cache(self) -> None:
+        """最適化キャッシュをクリア"""
+        self._operation_cache.clear()
+        self._metrics.clear()
+        self.logger.info("最適化キャッシュをクリアしました")
+
+    def _estimate_memory_usage(self, obj: Any) -> int:
+        """オブジェクトのメモリ使用量推定（簡易版）"""
+        try:
+            import sys
+
+            return sys.getsizeof(obj)
+        except Exception:
+            return 0
+
+    def _estimate_input_size(
+        self, args: tuple[Any, ...], kwargs: dict[str, Any]
+    ) -> int:
+        """入力サイズの推定"""
+        try:
+            total_size = 0
+            for arg in args:
+                if isinstance(arg, str):
+                    total_size += len(arg)
+                elif isinstance(arg, list) and arg and isinstance(arg[0], str):
+                    total_size += sum(len(line) for line in arg)
+            return total_size
+        except Exception:
+            return 0
